server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.0
mysql-connector-python==8.4.0
pycryptodome==3.20.0
bcrypt==4.2.0
PyJWT==2.9.0
slowapi==0.1.9
python-jose[cryptography]==3.4.0
secure==0.3.0
cryptography==43.0.0
email-validator>=2.0.0
